Add future monad
Why? Because I wanted generators to be able to execute synchronously OR
asynchronously depending on the implementation of the underlying effect.
The generator shouldn't know the difference, and I didn't feel like
making the interpreter care about that either.

`Future` provides a generic interface for effect composition in the
lower level of the interpreter.