Allow hierarchical context definitions
This is the commit that brings it all together. Now you can override
contexts on a hierarchy! That might not mean a lot when I say it that
way, so let me rephrase it: React hooks and context.

Let's say you want to use a different logging provider. If the logging
effects defined a console context, you could override the console at as
granular or global of a level as you prefer. Same thing for error
handling, configs, test doubles - whoops! That was supposed to be
a surprise. No spoilers :)